+++ mayan/apps/documents/column_widgets.py
@@ -7,8 +7,15 @@ from .settings import setting_thumbnail_height, setting_thumbnail_width
 class ThumbnailWidget(SourceColumnWidget):
     template_name = 'documents/widgets/thumbnail.html'
 
-    def disable_condition(self, instance):
-        return instance.is_in_trash
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.document = self.get_document()
+
+    def disable_condition(self):
+        return self.document.is_in_trash
+
+    def get_document(self):
+        return self.value
 
     def get_extra_context(self):
         transformation_resize = TransformationResize(
@@ -20,8 +27,8 @@ class ThumbnailWidget(SourceColumnWidget):
 
         return {
             # Disable the clickable link if the document is in the trash.
-            'disable_title_link': self.disable_condition(instance=self.value),
+            'disable_title_link': self.disable_condition(),
             'gallery_name': 'document_list',
-            'instance': self.value,
+            'instance': self.document,
             'transformation_instance_list': transformation_instance_list
         }