+++ mayan/apps/source_stored_files/column_widgets.py
@@ -4,5 +4,5 @@ from mayan.apps.documents.column_widgets import ThumbnailWidget
 class StoredFileThumbnailWidget(ThumbnailWidget):
     gallery_name = 'sources:stored_file_list'
 
-    def disable_condition(self, instance):
+    def disable_condition(self):
         return True