+++ HISTORY.rst
@@ -83,6 +83,8 @@
 
 - Remove KeyCloak from the Docker Compose file.
 - Support enabling or disabling file metadata driver per document type.
+- Pass the value to `SourceColumnWidget` subclasses during initialization
+  instead of during rendering.
 
 4.7.1 (2024-06-04)
 ==================